Add support for lists in the locale property
Since the accept-languages header can return a list of locales in preferred order, strong should be able to resolve the best locale from the list. This will allow the locale property to be set either as a string or an array (strings are converted to an array internally) and the locales will be searched in order for a matching key.